graduate the upscale setting
